requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
google-cloud-storage==2.10.0
python-magic==0.4.27
aiofiles==23.2.1